refactor: move pena link token slice into core application
