+++ backend/src/api/dependencies/use_cases.py
@@ -2,6 +2,9 @@ from auth.application.use_cases.login import LoginAdminUseCase, LoginUserUseCase
 from auth.infrastructure.repositories.sqlalchemy_auth_account_repository import (
     SqlAlchemyAuthAccountRepository,
 )
+from core.application.use_cases.generate_pena_link_token_usecase import (
+    GeneratePenaLinkTokenUseCase,
+)
 from core.application.use_cases.get_nationalities_usecase import GetNationalitiesUseCase
 from core.application.use_cases.get_pena_players_usecase import GetPenaPlayersUseCase
 from core.application.use_cases.get_penas_usecase import GetPenasUseCase
@@ -19,9 +22,6 @@ from core.application.use_cases.update_player_profile_usecase import (
     UpdatePlayerProfileUseCase,
 )
 from fastapi import Depends
-from persistence.application.use_cases.generate_pena_link_token_usecase import (
-    GeneratePenaLinkTokenUseCase,
-)
 from persistence.application.use_cases.link_user_to_pena_usecase import LinkUserToPenaUseCase
 from persistence.application.use_cases.manage_pena_accountability_usecase import (
     ManagePenaAccountabilityUseCase,
