+++ backend/src/api/interface/controller/v1/penas_controller.py
@@ -39,6 +39,9 @@ from auth.dependencies import (
     require_user,
 )
 from core.application.models import PenaLabelsUpdate, PenaProfileUpdate, PenasPage
+from core.application.use_cases.generate_pena_link_token_usecase import (
+    GeneratePenaLinkTokenUseCase,
+)
 from core.application.use_cases.get_penas_usecase import GetPenasUseCase
 from core.application.use_cases.manage_pena_labels_usecase import (
     ManagePenaLabelsUseCase,
@@ -48,7 +51,6 @@ from core.application.use_cases.manage_pena_profile_usecase import (
 )
 from fastapi import APIRouter, Depends, HTTPException, Query, status
 from persistence.application.use_cases import (
-    GeneratePenaLinkTokenUseCase,
     LinkUserToPenaUseCase,
     ManagePenaAccountabilityUseCase,
     PenaAccountabilityExpenseCreate,
