+++ backend/src/core/application/models/pena_link_models.py
@@ -1,0 +1,8 @@
+from dataclasses import dataclass
+
+
+@dataclass(frozen=True)
+class PenaLinkToken:
+    token: str
+    pena_guid: str
+    expires_at: int
