+++ backend/src/core/application/ports/__init__.py
@@ -1,5 +1,6 @@
 from .nationality_query_port import NationalityQueryPort
 from .pena_labels_port import PenaLabelsPort
+from .pena_link_port import PenaLinkPort
 from .pena_player_query_port import PenaPlayerQueryPort
 from .pena_profile_port import PenaProfilePort
 from .pena_query_port import PenaQueryPort
@@ -7,6 +8,7 @@ from .player_profile_port import PlayerProfilePort
 from .season_match_insights_port import SeasonMatchInsightsPort
 
 __all__ = [
+    "PenaLinkPort",
     "NationalityQueryPort",
     "PenaLabelsPort",
     "PenaPlayerQueryPort",
