+++ backend/src/core/application/ports/pena_link_port.py
@@ -1,0 +1,40 @@
+from dataclasses import dataclass
+from typing import Protocol
+
+
+@dataclass(frozen=True)
+class PenaLinkTokenResult:
+    token: str
+    pena_guid: str
+    expires_at: int
+
+
+class PenaNotManagedByAdminError(Exception):
+    pass
+
+
+class InvalidOrExpiredLinkTokenError(Exception):
+    pass
+
+
+class UserAlreadyLinkedToPenaError(Exception):
+    pass
+
+
+class UserPlayerNotFoundError(Exception):
+    pass
+
+
+class PenaLinkPort(Protocol):
+    def create_token_for_admin_pena(
+        self, *, admin_id: int, pena_guid: str, ttl_seconds: int
+    ) -> PenaLinkTokenResult: ...
+
+    def consume_token_for_user(
+        self,
+        *,
+        token: str,
+        account_id: int,
+        nickname: str | None,
+        position: str | None,
+    ) -> None: ...
