+++ backend/src/persistence/application/use_cases/generate_pena_link_token_usecase.py
@@ -1,35 +1,7 @@
-from dataclasses import dataclass
-
-from persistence.application.ports.pena_link_port import (
-    PenaLinkPort,
-    PenaNotManagedByAdminError,
+from core.application.models import PenaLinkToken
+from core.application.use_cases.generate_pena_link_token_usecase import (
+    GeneratePenaLinkTokenUseCase,
+    PenaAccessDeniedError,
 )
 
-
-class PenaAccessDeniedError(Exception):
-    pass
-
-
-@dataclass(frozen=True)
-class PenaLinkToken:
-    token: str
-    pena_guid: str
-    expires_at: int
-
-
-class GeneratePenaLinkTokenUseCase:
-    def __init__(self, repository: PenaLinkPort):
-        self.repository = repository
-
-    def execute(self, *, admin_id: int, pena_guid: str, ttl_seconds: int) -> PenaLinkToken:
-        try:
-            created = self.repository.create_token_for_admin_pena(
-                admin_id=admin_id, pena_guid=pena_guid, ttl_seconds=ttl_seconds
-            )
-        except PenaNotManagedByAdminError as exc:
-            raise PenaAccessDeniedError() from exc
-        return PenaLinkToken(
-            token=created.token,
-            pena_guid=created.pena_guid,
-            expires_at=created.expires_at,
-        )
+__all__ = ["GeneratePenaLinkTokenUseCase", "PenaAccessDeniedError", "PenaLinkToken"]
