+++ backend/src/persistence/infrastructure/repository/db/pena_link_repository.py
@@ -1,7 +1,7 @@
 import secrets
 import time
 
-from persistence.application.ports.pena_link_port import (
+from core.application.ports.pena_link_port import (
     InvalidOrExpiredLinkTokenError,
     PenaLinkPort,
     PenaLinkTokenResult,
@@ -9,8 +9,8 @@ from persistence.application.ports.pena_link_port import (
     UserAlreadyLinkedToPenaError,
     UserPlayerNotFoundError,
 )
+from core.domain.label_config import DEFAULT_ROLE_LABELS, pick_preferred_label
 from persistence.domain.entity import Pena, PenaLinkToken, PenaPlayer, PenaRole, Player
-from persistence.domain.label_config import DEFAULT_ROLE_LABELS, pick_preferred_label
 from sqlalchemy import delete, select
 from sqlalchemy.exc import IntegrityError
 from sqlalchemy.orm import Session
